Minor changes to request and response handling, created new JS file
