+++ capstone/src/main/java/com/google/sps/servlets/FormHandlerServlet.java
@@ -4,16 +4,8 @@ import java.io.BufferedReader;
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
-import com.google.appengine.api.datastore.PreparedQuery;
-import com.google.appengine.api.datastore.Query;
-import com.google.appengine.api.datastore.Query.SortDirection;
-import com.google.appengine.api.datastore.Key;
-import com.google.appengine.api.datastore.KeyFactory;
-import com.google.appengine.api.datastore.Entity;
 import com.google.gson.Gson;
 import com.google.sps.data.Form;
-import java.util.ArrayList;
-import java.util.List;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -40,5 +32,8 @@ public class FormHandlerServlet extends HttpServlet {
         Entity formEntity = form.toDatastoreEntity();
         DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
         datastore.put(formEntity);
+
+        response.setContentType("text/html");
+        response.setStatus(HttpServletResponse.SC_OK);
     }
 }
