+++ capstone/src/main/webapp/form.html
@@ -16,7 +16,7 @@
       <button onclick="callScriptFunction()">Create Form</button>
     </div>
 
-    <pre id="content" style="white-space: pre-wrap;"></pre>
+    <pre id="content"></pre>
     <script src = "form-loader.js"></script>
     <script async defer src="https://apis.google.com/js/api.js"
       onload="this.onload=function(){};handleClientLoad()"
